+++ example/init.lua
@@ -18,6 +18,11 @@ mng.package [[
 ]]
 mng.service_on("docker")
 mng.cmd("usermod -aG docker girvel")
--- TODO `winetricks dxvk` needs to be run once; consider --heavy flag? or --light?
+
+if not mng.cli_args.light then
+  mng.as_user("girvel", function()
+    mng.cmd("winetricks dxvk")
+  end)
+end
 
 mng.finish()